Renamed alias to slug and remove accents in slugs.
