+++ src/Logitar.Portal.UI/src/config/vee-validate/index.js
@@ -2,7 +2,6 @@ import Vue from 'vue'
 import { ValidationObserver, ValidationProvider, extend, localize } from 'vee-validate'
 import { alpha_num, confirmed, email, max, max_value, min, min_value, required } from 'vee-validate/dist/rules'
 
-import './rules/alias'
 import './rules/identifier'
 import './rules/purpose'
 import './rules/requireDigit'
@@ -10,6 +9,7 @@ import './rules/requireLowercase'
 import './rules/requireNonAlphanumeric'
 import './rules/requireUppercase'
 import './rules/requiredUniqueChars'
+import './rules/slug'
 import './rules/url'
 import './rules/username'
