+++ src/Logitar.Portal.UI/src/config/vee-validate/rules/slug.js
@@ -1,7 +1,7 @@
 import { extend } from 'vee-validate'
 import { isLetterOrDigit } from '@/helpers/stringUtils'
 
-extend('alias', {
+extend('slug', {
   validate(value) {
     return typeof value === 'string' && value.split('-').every(word => word.length && [...word].every(isLetterOrDigit))
   }
