+++ src/Logitar.Portal.UI/src/helpers/stringUtils.js
@@ -13,3 +13,45 @@ export function isLetter(c) {
 export function isLetterOrDigit(c) {
   return typeof c === 'string' && (isLetter(c) || isDigit(c))
 }
+
+export function shortify(text, length) {
+  return text?.length > length ? text.substring(0, length - 1) + '…' : text
+}
+
+export function slugify(value) {
+  value = value ?? ''
+  const words = []
+  let word = ''
+  for (let i = 0; i < value.length; i++) {
+    const c = value[i]
+    if (isLetterOrDigit(c)) {
+      word += c
+    } else if (word.length) {
+      words.push(word)
+      word = ''
+    }
+  }
+  if (word.length) {
+    words.push(word)
+  }
+  return unaccent(words.join('-').toLowerCase())
+}
+
+export function unaccent(input) {
+  const accents = {
+    à: 'a',
+    â: 'a',
+    ç: 'c',
+    è: 'e',
+    é: 'e',
+    ê: 'e',
+    ë: 'e',
+    î: 'i',
+    ï: 'i',
+    ô: 'o',
+    ù: 'u',
+    û: 'u',
+    ü: 'u'
+  }
+  return [...input].map(c => (c.toUpperCase() === c ? (accents[c] ?? c).toUpperCase() : accents[c] ?? c)).join('')
+}
